scaffold: two-pass lfs loading + etc.
Also includes:

- Changed to synchronous NTP sync. This was backgrounded before,
  together with the MQTT connect. However, this wouldn't be much faster,
  and often would fail (because of a lack of true parallism,
  presumably).
- Saving/displaying the last-modified value for the LFS image, when it
  was retrieved.
- Refactor into generate_status_json() for both connect and ping
  responses.
- "myname" now is no longer configurable. It was never used, and
  unnecessary complexity arose from it.

The main change, however, is that the scaffold now uses http_ota.lua
(and not HTTP_OTA.lua from nodemcu-firmware), which I've modified in a
couple of ways to allow a two-pass LFS load. This means:

- http_ota.lua loads the file into lfs_inc.img (for "incoming").
- If that succeeds—and where HTTP_OTA.lua would have progressed to
  node.flashreload()—the file is simply left there, and the device is
  rebooted.
- If init.lua—before attempting any LFS loads—establishes that a
  "lfs_inc.img" is present, it moves it into "lfs_flashed.img" and then
  runs node.flashreload() upon that file.
- If this fails, init.lua puts the resultant error message into
  lfs.errormsg.
- Before all of this, init.lua clears lfs.errormsg and lfs_flashed.img.

The node.task.post() approach to flashreload'ing seems pretty reliable.
However, I have had issues with flashreload()—instead of doing the
flashing and then restarting—would cause a watchdog timeout.

Unfortunately, there is hardly any way to establish which scenario
happened, owing the non-returning nature of flashreload(). But a slim
(?) probability of flashes not carrying through exists, through this
system.